requirements.txt
Flask==3.0.3
Flask_APScheduler==1.13.1
Flask_PyMongo==2.3.0
numpy==2.0.2
selenium==4.26.1
torch==2.5.1
transformers==4.46.2
flask_cors==5.0.0
python-dotenv==1.0.1
